review-management-service/requirements.txt
flask
boto3
flask-cors
python-dotenv
psycopg2